Move new "should join persons" check to `TrendsEventQuery`
