+++ ee/clickhouse/queries/test/test_event_query.py
@@ -10,7 +10,7 @@ from posthog.models.filters import Filter
 from posthog.models.group.util import create_group
 from posthog.models.group_type_mapping import GroupTypeMapping
 from posthog.models.person import Person
-from posthog.queries.trends.trend_event_query import TrendsEventQuery
+from posthog.queries.trends.trends_event_query import TrendsEventQuery
 from posthog.test.base import (
     APIBaseTest,
     ClickhouseTestMixin,
